fix: fix import, move view
