+++ src/views/_proxied-dot-io/product-docs-landing/vault-get-started-diagram.png
